Extract theme related functions (previously in style.rs) into a new theme mod and abstract into `trait Theme`
